+++ Core/StitchCoreSDK/Tests/StitchCoreSDKTests/Internal/Net/StitchRequestClientUnitTests.swift
@@ -70,7 +70,7 @@ class StitchRequestClientUnitTests: StitchXCTestCase {
 
         XCTAssertEqual(response.statusCode, 200)
 
-        let expected = ["hello": "world", "a": 42] as Document
+        let expected = ["hello": "world", "a": 42 as Int32] as Document
         XCTAssertEqual(expected, try Document.init(fromJSON: response.body!))
 
         transport.mockRoundTrip.clearStubs()
@@ -177,7 +177,7 @@ class StitchRequestClientUnitTests: StitchXCTestCase {
         )
 
         let path = "/path"
-        let document: Document = ["my": 24]
+        let document: Document = ["my": 24 as Int32]
         let builder = StitchDocRequestBuilder()
             .with(path: path)
             .with(method: .get)
@@ -230,7 +230,7 @@ class StitchRequestClientUnitTests: StitchXCTestCase {
 
         XCTAssertEqual(response.statusCode, 200)
 
-        let expected = ["hello": "world", "a": 42] as Document
+        let expected = ["hello": "world", "a": 42 as Int32] as Document
         XCTAssertEqual(expected, try Document.init(fromJSON: response.body!))
 
         transport.mockRoundTrip.clearStubs()
